Changed output formats
* applied some changes from last commit from synthseg
* added nu_correction script for testing
* save hemis-seg partition in float32, because of rounding issues with uint8
* skip bias correction in CAT_VolAmap
* save skull-stripped image as float32 because of rounding issues
* changed bias correction scheme

+++ scripts/partition_hemispheres.py
@@ -49,5 +49,5 @@ for j in [0, 1]:
     crop_idx = utils.bbox_volume(hemi > 1, pad=5)
     hemi, aff_hemi = edit_volumes.crop_volume_with_idx(hemi, crop_idx, aff=aff_amap, n_dims=3, return_copy=False)
 
-    tools.save_volume(hemi, aff_hemi, h_amap, hemi_name, dtype='uint8')
+    tools.save_volume(hemi, aff_hemi, h_amap, hemi_name, dtype='float32')
